+++ game_patch/os/console.h
@@ -17,6 +17,7 @@ void console_apply_patches();
 void console_register_command(rf::console::Command* cmd);
 rf::Player* find_best_matching_player(const char* name);
 void console_start_server_log();
+void console_run_script(const char* filename);
 
 class DcInvalidArgTypeError : public std::exception
 {};